feat: path 재조정 및 접기 버튼, css 사소한 변경
