+++ src/app/customer-support/privacy-policy/page.tsx
