+++ src/data/policy/service.mdx
@@ -27,22 +27,22 @@
 5. 회원은 변경된 약관 사항에 동의하지 않으면 서비스 이용을 중단하고 이용 계약을 해지할 수 있습니다.
 6. 개별 상품 판매약관에 대한 동의 및 약관개정절차 등 제반 사항은 해당 판매자가 정한 바에 따릅니다.
 <br />
-## **제4조 (약관의 해석 및 약관 외 준칙)**
+## 제4조 (약관의 해석 및 약관 외 준칙)
 <br />
 1. 회사는 필요한 경우 제7조에서 규정하는 서비스의 상세 내용 중 개별 서비스에 관하여 적용될 사항을 위해 운영정책, 이용지침 등(이하 통칭하여 "운영정책 등" 이라 합니다)을 정하여 운영할 수 있습니다.
 2. 본 약관에 규정되지 않은 사항에 대해서는 개별 운영정책, 관련 법령과 일반 상관례에 의합니다.
 3. 회원은 본 약관 및 운영정책 등의 내용에 변경이 있는지 여부를 확인하여야 합니다. 회사는 회원이 이를 확인하지 않아 발생한 손해에 대해서는 책임을 부담하지 않습니다.
 <br />
-# **제2장 서비스의 이용**
+# 제2장 서비스의 이용
 <br />
-## **제5조 (이용신청)**
+## 제5조 (이용신청)
 <br />
 1. 서비스 이용계약은 이용 희망자가 본 약관에 동의하고 가입절차를 완료하여 서비스를 이용함으로써 체결됩니다.
 2. 회사는 서비스 가입 희망자에게 본인확인을 위해 휴대 단말 식별 정보 및 본인확인 인증 절차(이하 "본인인증 절차")를 요구할 수 있습니다. 이 경우, 가입 희망자는 해당 인증을 수행하여 가입 절차를 완료해야 합니다.
 3. 회원의 관리 소홀로 타인이 회원 본인 명의의 인증 절차를 수행하거나, 회원이 타인의 명의를 도용하여 인증을 수행하는 경우, 이로 인해 발생하는 불이익이나 손해에 대해 회사는 책임지지 않습니다.
 4. 회원은 서비스 이용을 위해 등록 시 현재의 사실과 일치하는 완전한 정보(이하 "등록정보")를 제공해야 하며, 등록정보에 변경사항이 발생할 경우 즉시 갱신해야 합니다. 회원이 제공한 등록정보 및 갱신한 등록정보가 부정확할 경우, 회사는 서비스 이용을 제한하거나 중단할 수 있습니다.
 <br />
-## **제6조 (이용신청에 대한 승낙 유보 및 서비스 이용의 제한)**
+## 제6조 (이용신청에 대한 승낙 유보 및 서비스 이용의 제한)
 <br />
 1. 제5조에 의한 가입 신청자의 신청에 대하여 회사는 원칙적으로 서비스 이용을 승낙합니다. 단, 회사는 다음 각 호에 해당하는 신청에 대하여 그 사유가 해소될 때까지 승낙을 유보하거나 거부할 수 있습니다.
    ① 타인의 명의로 이용신청을 한 경우
@@ -59,7 +59,7 @@
 3. 회원이 본 이용약관에 동의하고 핸디버스를 이용한 서비스에 가입하는 것은, 핸디버스에서 제공하는 서비스를 회사가 제공하는 방식과 내용으로 이용하며 회사가 정한 본 약관 및 운영정책 등을 준수한다는 의사를 표시한 것으로 간주됩니다.
 4. 회사는 서비스의 품질 향상 및 기타 필요한 경우, 회원을 등급별로 구분하여 이용시간, 이용횟수, 서비스 메뉴 등 필요한 사항을 세분화하여 그 이용에 차등을 둘 수 있습니다.
 <br />
-## **제7조 (서비스의 내용)**
+## 제7조 (서비스의 내용)
 <br />
 1. 회사가 핸디버스를 통해 제공하는 서비스는 아래와 같습니다.
 <br />
@@ -69,12 +69,12 @@
 <br />
 1. 회사는 직접 판매자(운송사업자)가 제공하는 다양한 상품 및 서비스를 핸디버스를 통해서 제공할 수 있습니다. 해당 서비스 이용과 관련하여서는 직접 판매자의 이용약관을 따르고 회사는 회원이 해당 이용약관을 확인할 수 있도록 직접 판매자의 홈페이지 등에 게시되어 있는 이용약관을 연결합니다. 또한 회사는 회원이 직접 판매자의 서비스를 이용하기 전에 핸디버스를 이용한 서비스의 화면을 통하여 판매사의 이용요금 및 수수료 부과 기준, 수수료 금액 등을 안내합니다.
 <br />
-## **제8조 (서비스의 이용 개시)**
+## 제8조 (서비스의 이용 개시)
 <br />
 1. 회사는 회원의 이용 신청을 승낙한 때부터 서비스를 개시합니다. 단, 일부 서비스의 경우에는 지정된 일자부터 서비스를 개시합니다.
 2. 회사의 업무상 또는 기술상의 장애로 인하여 서비스를 개시하지 못하는 경우에는 서비스 모바일 어플리케이션 또는 서비스 웹사이트에 공시하거나 회원에게 이를 통지합니다.
 <br />
-## **제9조 (서비스의 이용시간)**
+## 제9조 (서비스의 이용시간)
 <br />
 1. 서비스는 연중무휴 1일 24시간 제공을 원칙으로 하며, 구체적인 종류와 세부내용은 서비스 내 사용방법 안내를 따릅니다. 다만, 회사의 업무상 또는 기술상의 이유로 서비스가 일시 중지될 수 있으며, 운영상의 목적으로 회사가 정한 기간에도 서비스가 일시 중지될 수 있습니다. 이러한 경우 회사는 사전 또는 사후에 이를 공지합니다.
 2. 회사는 기술적 사양의 변경이나 기타 불가피한 여건 또는 사정이 있는 경우 서비스의 내용을 변경할 수 있습니다. 이 경우 회사는 변경된 서비스의 내용 및 제공 일자를 명시하여 이용자에게 공지합니다. 이용자는 변경된 서비스의 내용을 확인할 수 있습니다.
@@ -90,7 +90,7 @@
 ① 이용자는 신용카드, 계좌 이체 또는 기타 회사가 추가로 정의하는 결제수단을 통해 셔틀을 예약합니다.
 ② 서비스 내에서 좌석 예약 시 회사는 이용자에게 소정의 예약 수수료를 청구할 수 있으며, 효율적인 예약과 무분별한 예약 후 취소를 방지하기 위해, 예약 취소시에는 아래 정책에 따라 결제 금액의 일부 또는 전부를 위약금으로 공제합니다.
 <br />
-**[취소 및 환불 안내]**
+[취소 및 환불 안내]
 <br />
 탑승 시간을 기준으로 환불 신청 시점에 따라 수수료가 발생합니다.
 <br />
@@ -100,7 +100,7 @@
 <br />
 시간대는 한국 기준으로 적용됩니다.
 <br />
-| **취소시점** | **취소 수수료** |
+| 취소시점 | 취소 수수료 |
 |--------|-------|
 | 예약 24시간 이내| 수수료없음 |
 | 탑승 D-8 23:59까지| 수수료 없음 |
@@ -114,7 +114,7 @@
 <br />
 이용자 및 회사는 이용자의 서비스 사용 및 회원 가입에 따른 본 약관 상의 지위 또는 권리, 의무의 전부 또는 일부를 제3자에게 양도, 위임하거나 담보제공 등의 목적으로 처분할 수 없습니다.
 <br />
-## **제13조 (서비스의 변경 및 중단)**
+## 제13조 (서비스의 변경 및 중단)
 <br />
 1. 회사는 다음 각 호에 해당하는 경우 서비스의 전부 또는 일부를 제한하거나 중단할 수 있습니다.
    ① 서비스용 설비의 보수 등 공사로 인한 부득이한 경우
@@ -125,7 +125,7 @@
 2. 제1항에 의한 서비스 중단의 경우에는 회사가 제16조에서 정한 방법으로 회원에게 통지합니다. 다만, 회사가 통제할 수 없는 사유로 인한 서비스의 중단(회사의 고의, 과실이 없는 디스크 장애, 시스템다운 등)으로 인하여 사전 통지가 불가능한 경우에는 그러하지 아니합니다.
 3. 회사는 무료로 제공되는 서비스의 변경, 중단으로 발생하는 문제에 대해서 회사에 고의, 중과실이 없는 한 어떠한 책임도 지지 않습니다.
 <br />
-## **제14조 (이용계약해지 및 이용제한)**
+## 제14조 (이용계약해지 및 이용제한)
 <br />
 1. 회원은 언제든지 회사에 이용계약 해지의 의사를 통지함으로써 이용계약을 해지할 수 있습니다. 다만, 회원은 해지의사를 통지하기 전에 현재 진행 중인 서비스의 거래를 완료하거나 철회 또는 취소하여야 합니다.
 2. 회사는 제1항에 따른 이용계약 해지로 인해 발생할 수 있는 불이익 등에 대하여 귀책사유가 없으면 책임을 부담하지 않습니다. 회사가 이용계약에 기초하여 회원에게 제공한 혜택은 이용계약과 함께 종료됩니다.
@@ -137,7 +137,7 @@
 8. 회사가 본 조항에 따라 서비스 이용을 제한하거나 이용계약을 해지하는 경우, 회사는 사전에 회원에게 통지합니다. 단, 부득이한 사유로 인하여 사전 통지가 불가능한 경우에는 그 사유가 종료된 후 즉시 통지합니다.
 9. 회원은 전항의 통지일로부터 10일 이내에 항변 및 소명을 할 수 있고, 해당 항변 및 소명이 정당한 경우 회사는 서비스 이용 제한 또는 이용계약 해지를 취소할 수 있습니다. 단, 회원이 기한 내 항변 또는 소명을 하지 않은 경우, 서비스 이용 제한 또는 이용계약 해지에 동의한 것으로 간주합니다.
 <br />
-# **제3장 회사와 이용자의 권리 및 의무**
+# 제3장 회사와 이용자의 권리 및 의무
 <br />
 ## 제15조 (회사의 의무)
 <br />
@@ -151,21 +151,21 @@
 ② 이용자는 본 약관 제6조 제1항에 기재된 행위를 하여서는 안됩니다.
 ③ 이용자는 관계 법령 및 이 약관의 규정 또는 이용 안내 등 회사가 통지하는 사항을 준수 하여야 하며, 기타 회사의 업무에 방해가 되는 행위를 해서는 안됩니다.
 <br />
-# **제4장 기타**
+# 제4장 기타
 <br />
 ## 제17조 (이용자에 대한 통지)
 <br />
 ① 회사가 회원에 대하여 통지를 하는 경우 회원이 회원 가입 시 제공한 이메일 주소 또는 이동전화 번호를 통해 통지할 수 있습니다.
 ② 불특정 다수의 이용자에 대한 통지의 경우 어플리케이션 또는 홈페이지 내의 공지사항에 게시함으로써 개별 통지에 갈음할 수 있습니다.
 <br />
-## **제18조 (개인정보 등의 변경 및 보호)**
+## 제18조 (개인정보 등의 변경 및 보호)
 <br />
 1. 회사는 회원이 안전하게 서비스를 이용할 수 있도록 회원의 개인정보, 위치정보, 신용정보 등(이하 “개인정보 등"이라 함)을 보호하기 위하여 관련법령에 따라 보안 시스템 등을 갖추고 개인정보 등 보호 정책을 실시합니다.
 2. 회사는 회원의 개인정보 등을 관련법령상 필요한 경우 등을 제외하고 서비스를 제공하기 위한 목적 이외의 용도로 사용하거나 회원의 동의 없이 제3자에게 제공하지 않습니다. 단, 다음의 경우는 예외로 합니다.
 <br />
-   ① 서비스 제공에 관한 계약을 이행하기 위한 것으로 제휴사에게 거래 내역, 정산 내역 제공등에 필요한 최소한의 회원정보(성명, 연락처 등)를 제공하는 경우
+① 서비스 제공에 관한 계약을 이행하기 위한 것으로 제휴사에게 거래 내역, 정산 내역 제공등에 필요한 최소한의 회원정보(성명, 연락처 등)를 제공하는 경우
 <br />
-   ② 법령에서 회원의 개인정보 이용과 제3자에 대한 정보제공을 허용하고 있는 경우
+② 법령에서 회원의 개인정보 이용과 제3자에 대한 정보제공을 허용하고 있는 경우
 <br />
 3. 회사는 수집된 개인정보 등의 취급과 관리 등의 업무를 스스로 수행함을 원칙으로 합니다. 다만, 회사는 필요한 경우 회원의 동의를 받거나 통지를 한 후 위 업무의 일부 또는 전부를 회사가 선정한 회사에 위탁할 수 있습니다.
 4. 회원은 이용신청 시 기재한 사항이 변경되었을 경우에는 즉시 변경사항을 최신의 정보로 수정하여야 합니다. 단, 회사가 정한 정보 항목은 회사의 승인을 받아 수정할 수 있습니다.
@@ -198,7 +198,7 @@
 <br />
 본 약관에 명시되지 않은 사항에 대해서는 「전자금융거래법」, 「정보통신망 이용촉진 및 정보보호 등에 관한 법률」 등 관계 법령과 회사가 정한 서비스의 세부 이용 지침 등에 따릅니다.
 <br />
-## **제23조 (면책사항)**
+## 제23조 (면책사항)
 <br />
 1. 회사는 핸디버스를 이용한 수요응답형 모빌리티 서비스를 이용하기 위한 플랫폼을 운영•관리•제공할 뿐 회원과 판매자(운송서비스사업자 등 포함)들을 대리하지는 않습니다. 또한 판매자의 업무범위에서 발생되는 서비스의 이용 장애 및 각종 사고 등에 대하여, 회사의 귀책사유가 없는 한 책임을 지지 않습니다. 판매사와 회사의 업무범위는 양사간의 계약 내용을 기반으로 합니다.
 2. 회사는 천재지변 또는 이에 준하는 불가항력으로 인하여 서비스를 제공할 수 없는 경우에는 서비스 제공에 관한 회사의 귀책사유가 없습니다.
@@ -206,6 +206,6 @@
 4. 회사는 회원이 서비스를 이용하여 기대하는 수익을 상실한 것에 대하여는 무관하며 서비스를 통하여 얻은 2차적인 자료로 인한 손해 등에 대하여도 책임을 지지 않습니다. 회사는 회원이 서비스에 게재한 정보, 자료, 사실의 신뢰도 및 정확성 등 내용과는 무관합니다.
 5. 회사는 회원 상호간 또는 회원과 제3자 상호간에 서비스를 매개로 발생한 분쟁에 대해서는 개입하지 않습니다.
 <br />
-**부칙**
+부칙
 <br />
 본 약관은 2024년 10월 14일부터 적용됩니다.
